flask_backend/requirements.txt
flask
flask-cors
openai
numpy
pandas
newsapi-python
google-generativeai